Start of project, defines, simple functions
Added types for use with coordinates, but as a floating number as well
as in the arcdeg, arcmin and arcsec format.
Created functions to convert from dms (arcdeg, arcmin, arcsec) to a
floating number in degrees.
Defined the standard of which degrees are signed depending on N/S and
E/W.
Created a rudimentary coordinate structure for use in waypoints and
current location.