New distance calculations, reduced warnings
Added new methods of calculating distance between coordinates as well as
removed sources of compiler warnings due to lack of typecasting. If the
floatDegree remains a 32 bit float it will now typecast all doubles to
this after calculations. If floatDegree is later changed to double this
should cause no issues.
Added distance function by the spherical cosine law method.
Added distance function by the equirectangular approximation method.
Created a simple test for the above functions to compare the results to
the one that is calculated using the haversine version of the spherical
cosine law. Ideally they should net the same result but might differ at
extremes. Big todo to test these algorithms for a wide range of
coordinate sets.