Continued nav system and matlab scripts
Started thinking about the nav state of the system of which can be used
to keep track of what the system currently needs to be doing. For
example when to look for next WP, when on the way to a WP etc.

Started matlab implementation of haversine and great circle distance
formulae. Need to implement higher accuracy ellipsoid formulae and
compare the accuracy of the two.