Bugfix with conversion of negative degrees.
Refactoring for UART write function to account for overflows and return an error.

Wrote NavTester to accept a csv file as input, hold the WP static and output a csv file with current coord, WP coord and distance between them.

Added Matlab script to read the test data from NavTester and compare the calculated values to those produced by Matlab using the WSG84 model of the Earth.