Changed PSoC ISR definition and helper functions to navisr.h/c.
Need to test if this way of doing ISR works, if not there might be a problem with the initial assumption on how UART interrupts are triggered when running with a UART block in PSoC Creator with less or equal to 4 B of buffer (Hardware buffering).

Changed the ISR trigger to "RISING EDGE" instead of "DERIVED". If this does not work try the "LEVEL" tied to the hardware FIFO fill state.