UART Rx ISR now works on the PSoC.
Issue was a user problem where a call to enable the UART was  not executed before trying to access the ISR signals.

Next step is to fully implement the FIFO buffer to detect when a full string is available in the buffer, extract the string and then process the data it contains.