Corrected typo in config.h.
Added TODO tasks for Visual Studio.

UART bug found when attempting to echo long strings sent from a computer via the Arduino and to the PSoC.
Attempt to echo just with the Arduino and see if the bug can be caused by the Arduino. If not see if it might be a problem with the method chosen for echoing on the PSoC.