Continued the Python testing interface.
The TestInterface can now read csv files, open a serial connection, read rows from the input csv file and format this to byte strings that can be sent over serial. The interface then waits for a response from the PSoC which when received is converted from byte strings into UTF-8 strings which are written to an output file.