Fixed bug with sprintf() on PSoC.
First of all had to add linker flag "-u _printf_float" since this type of functionality is disabled by default.
Ref: http://www.cypress.com/?app=forum&id=2492&rID=86619

Secondly since this is using precompiled headers there are hidden mallocs that take use of the heap. The default heap size is set to 0x80 Byte, leading to undefined behaviour when using more than this. Setting the heap size to 0x2000 Byte fixed the issue and the serial communication depending on sprintf() now works.