Moved all source/header files. Added file handling.
Major reallocation of source and header files in order to group them in
a distinct location.

PSoC project updated to new file locations but VS2013 project has yet to
be done.

Added some basic waypoint handling routines in wphandler.h/c that will
serve as an abstraction to the WP management for the rest of the system.
Yet to test the current implementation, should be current main priority.

Also changed the file headers and data block headers in order to allow
for future development of data structures that contain additional data
while still allowing an older system to operate with new types of files.