Testing SD card write / read between systems.
The enumerated types for PSoC projects are defaulting to the lowest
bytes of data needed to store all types. This causes a problem with the
enumerated types on other systems where they may or may not default to
1, 2 or 4 B.

In order to avoid ambiguity in the size of enumerated types a maximum
value of 2^32 is defined as the last value in the list, forcing the PSoC
to treat the enumerated as an integer of 4 B.

The WaypointCreator project can now write waypoint files that the PSoC
can read, making the next step a proper implementation of the storage
and retrieval of these files.