Added file functions and test functions.
Working on testing for WPHandler which should be the next thing to
finish.

Added NAV_ftell() interface so we can get the size of files.

Refactor target: The return values for NAV file interfaces are number of
items written - NOT the size of them. Refactor the counter variables so
that they are logical and represent the number of items written / read
and not the size of the items.